requirements.txt
fastapi
uvicorn
python-multipart
setuptools
webauthn
argon2-cffi
sqlalchemy
icecream
pydantic
pydantic[email]
python-dotenv
psycopg2-binary
geopy
redis
